refactor(auth): split registration form into SSR and client components
